Az bot: add templates/ to setup.py, reorganize tests' helper classes
