--- src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/tests/latest/tools/__init__.py
@@ -1,7 +1,0 @@
-# --------------------------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for license information.
-# --------------------------------------------------------------------------------------------
-
-import pkg_resources
-pkg_resources.declare_namespace(__name__)
